refactor(homepage): updates fetch logic + conditional path

+++ apps/frontend/pages/index.tsx
@@ -1,7 +1,9 @@
 /* eslint-disable no-underscore-dangle */
 import { useEffect } from 'react';
 
-import type { GetStaticProps, NextPage } from 'next';
+import type { NextPage } from 'next';
+import Router from 'next/router';
+import { useAuthContext } from 'src/hooks/useAuthContext';
 import { useWorkoutsContext } from 'src/hooks/useWorkoutsContext';
 
 import { Navbar, WorkoutForm } from '@components';
@@ -16,18 +18,43 @@ export interface Workout {
   updatedAt: string;
 }
 
-const Home: NextPage = ({ allWorkouts }: any) => {
+const Home: NextPage = () => {
   const { workouts, dispatch }: any = useWorkoutsContext();
+  const { user }: any = useAuthContext();
 
   useEffect(() => {
-    dispatch({ type: 'SET_WORKOUTS', payload: allWorkouts });
-  }, []);
+    const fetchWorkouts = async () => {
+      const req = await fetch('http://localhost:4000/api/workouts', {
+        headers: {
+          Authorization: `Bearer ${user.token}`,
+        },
+      });
+      const allWorkouts = await req.json();
+
+      if (req.ok) {
+        dispatch({ type: 'SET_WORKOUTS', payload: allWorkouts });
+      }
+    };
+
+    if (user) {
+      fetchWorkouts();
+    } else {
+      Router.push('/login');
+    }
+  }, [dispatch, user]);
 
   const handleClick = async (workout: Workout) => {
+    if (!user) {
+      return;
+    }
+
     const res = await fetch(
       `http://localhost:4000/api/workouts/${workout._id}`,
       {
         method: 'DELETE',
+        headers: {
+          Authorization: `Bearer ${user.token}`,
+        },
       }
     );
     const json = await res.json();
@@ -39,23 +66,27 @@ const Home: NextPage = ({ allWorkouts }: any) => {
 
   return (
     <div>
-      <Navbar />
-      {workouts && (
+      {user && (
         <>
-          <Grid>
-            <Stack display="flex" vertical>
-              {workouts.map((workout: Workout) => {
-                return (
-                  <Workouts
-                    key={workout._id}
-                    data={workout}
-                    handleClick={handleClick}
-                  />
-                );
-              })}
-            </Stack>
-            <WorkoutForm />
-          </Grid>
+          <Navbar />
+          {workouts && (
+            <>
+              <Grid>
+                <Stack display="flex" vertical>
+                  {workouts.map((workout: Workout) => {
+                    return (
+                      <Workouts
+                        key={workout._id}
+                        data={workout}
+                        handleClick={handleClick}
+                      />
+                    );
+                  })}
+                </Stack>
+                <WorkoutForm />
+              </Grid>
+            </>
+          )}
         </>
       )}
     </div>
@@ -63,14 +94,3 @@ const Home: NextPage = ({ allWorkouts }: any) => {
 };
 
 export default Home;
-
-export const getStaticProps: GetStaticProps = async () => {
-  const req = await fetch('http://localhost:4000/api/workouts');
-  const allWorkouts = await req.json();
-
-  return {
-    props: {
-      allWorkouts,
-    },
-  };
-};
